Replace our join_table by a join_queue.
So fibo can work.
There is some stuff weird with is_joining which is : I can't get rid of
it. A fixme is in the source for that.

We should probably use spin_lock to become faster.
And also find a way to remove easily all debug printf.
And find a way to free all global variable when the main call return
instead of thread_exit.
In short, see the TODO list.